DB Fixes
git-svn-id: http://cyanogen-updater.googlecode.com/svn/trunk@651 fa1c2e9a-81d2-11de-92f0-6bbadbc12360

+++ src/cmupdaterapp/utils/Preferences.java
@@ -169,7 +169,7 @@ public class Preferences extends Activity {
         DbAdapter themeListDb = new DbAdapter(showDebugOutput);
         if (showDebugOutput) Log.d(TAG, "Opening Database");
         themeListDb.open();
-        int count = themeListDb.getThemeCount();
+        long count = themeListDb.getThemeCount();
         themeListDb.close();
         return count > 0;
     }
